+++ src/main/java/com/finance/loans/infra/rest/ClientController.java
@@ -19,8 +19,8 @@ import com.finance.loans.infra.rest.dtos.ClientResponse;
 import com.finance.loans.infra.rest.dtos.LoanPostponeResponse;
 import com.finance.loans.infra.rest.dtos.LoanRequest;
 import com.finance.loans.infra.rest.dtos.LoanResponse;
-import com.finance.loans.repositories.Client;
-import com.finance.loans.repositories.Loan;
+import com.finance.loans.repositories.entities.Client;
+import com.finance.loans.repositories.entities.Loan;
 import com.finance.loans.repositories.mapper.ClientMapper;
 import com.finance.loans.repositories.mapper.LoanMapper;
 import com.finance.loans.repositories.mapper.LoanPostponeMapper;
