+++ src/main/java/com/finance/loans/repositories/ClientRepository.java
@@ -5,6 +5,8 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.finance.loans.repositories.entities.Client;
+
 @Repository
 public interface ClientRepository extends JpaRepository<Client, String> {
