+++ src/main/java/com/finance/loans/repositories/LoanRepository.java
@@ -5,6 +5,8 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.finance.loans.repositories.entities.Loan;
+
 @Repository
 public interface LoanRepository extends JpaRepository<Loan, Integer> {
