+++ src/main/java/com/finance/loans/repositories/mapper/LoanPostponeMapper.java
@@ -3,7 +3,7 @@ package com.finance.loans.repositories.mapper;
 import org.mapstruct.Mapper;
 
 import com.finance.loans.infra.rest.dtos.LoanPostponeResponse;
-import com.finance.loans.repositories.LoanPostpone;
+import com.finance.loans.repositories.entities.LoanPostpone;
 
 @Mapper(componentModel = "spring")
 public interface LoanPostponeMapper {
